refactor(server): extract run config freshness from heartbeat.ts
Moves the session/workspace config fingerprint metadata builders and
freshness resolution (session categories, workspace categories, model
change detection, session metadata attach/strip helpers) into
heartbeat-run-config-freshness.ts, following the heartbeat-<domain>
sibling pattern. heartbeat.ts re-exports the previously public symbols.
No behavior change.
